sqlite: separate object readers and selected scope execution
