+++ storage/sqlite/src/objects/index.ts
@@ -1,0 +1,2 @@
+export type { SqliteObjectStorageOptions } from "./storage"
+export { SqliteObjectStorage } from "./storage"
